+++ common/WebPShopUtils.cpp
@@ -13,7 +13,9 @@
 // limitations under the License.
 
 #include <time.h>
+
 #include <string>
+
 #include "PIUSuites.h"
 #include "WebPShop.h"